Rename: [py_]transaction_executor.rs
It is already a non-pyo3 struct, but the file was kept with a `py_`
prefix for easy review.

This commit is just a rename, no other changes.
