+++ crates/native_blockifier/src/lib.rs
@@ -6,9 +6,9 @@ pub mod py_state_diff;
 pub mod py_test_utils;
 pub mod py_transaction;
 pub mod py_transaction_execution_info;
-pub mod py_transaction_executor;
 pub mod py_utils;
 pub mod storage;
+pub mod transaction_executor;
 
 use errors::add_py_exceptions;
 use py_block_executor::PyBlockExecutor;
